Move AccountsSelector to the separate component and add dynamic loading
