+++ src/components/feeds-selector/selector.jsx
@@ -1,6 +1,5 @@
 import cn from 'classnames';
 import { useCallback, useEffect, useMemo } from 'react';
-import CreatableSelect from 'react-select/creatable';
 import { ButtonLink } from '../button-link';
 import { useBool } from '../hooks/bool';
 import { DisplayOption, useSelectedOptions } from './options';
@@ -18,6 +17,7 @@ import {
 } from './constants';
 import styles from './selector.module.scss';
 import { SelectorError } from './error';
+import { AccountsSelector } from './accounts-selector';
 
 export function Selector({
   className,
@@ -110,24 +110,16 @@ export function Selector({
             </ButtonLink>
           </>
         ) : (
-          <CreatableSelect
-            className={styles['selector']}
-            theme={selTheme}
+          <AccountsSelector
             placeholder={
               isDirect || mode === CREATE_DIRECT ? 'Select recipients...' : 'Select feeds...'
             }
-            isMulti={true}
-            isClearable={false}
             autoFocus={startStatic}
-            closeMenuOnSelect={true}
             openMenuOnFocus={!isEditing(mode)}
             options={options}
             value={values}
-            styles={selStyles}
             onChange={handleChange}
-            formatOptionLabel={formatOptionLabel}
             formatCreateLabel={formatCreateLabel}
-            isValidNewOption={isValidNewOption}
           />
         )}
       </div>
@@ -141,68 +133,6 @@ export function Selector({
   );
 }
 
-const selStyles = {
-  control: (base, state) => ({
-    ...base,
-    boxShadow: 'none',
-    '&:hover': { borderColor: 'var(--selector-color-neutral60)' },
-    borderColor: state.isFocused ? 'var(--selector-color-neutral60)' : base.borderColor,
-  }),
-  multiValueLabel: (base, state) => {
-    const s = { ...base, color: 'currentColor' };
-    return state.data.isFixed ? { ...s, paddingRight: 6 } : s;
-  },
-  multiValueRemove: (base, state) => {
-    const s = {
-      ...base,
-      borderLeft: '1px solid var(--selector-color-primary50)',
-      marginLeft: '4px',
-    };
-    return state.data.isFixed ? { ...s, display: 'none' } : s;
-  },
-  indicatorSeparator: (base) => ({ ...base, display: 'none' }),
-  multiValue: (base) => ({
-    ...base,
-    backgroundColor: 'var(--selector-color-primary25)',
-    color: 'var(--selector-color-value)',
-    borderRadius: '2px',
-    border: '1px solid var(--selector-color-primary50)',
-  }),
-};
-
-// Only valid usernames are allowed
-function isValidNewOption(label) {
-  return /^[a-z\d]{3,25}$/i.test(label.trim());
-}
-
 function formatCreateLabel(label) {
   return `Send direct message to @${label}`;
 }
-
-function formatOptionLabel(option) {
-  return <DisplayOption option={option} />;
-}
-
-const selTheme = (theme) => ({
-  ...theme,
-  borderRadius: 0,
-  colors: {
-    danger: 'var(--selector-color-danger)',
-    dangerLight: 'var(--selector-color-danger-light)',
-    neutral0: 'var(--selector-color-neutral0)',
-    neutral5: 'var(--selector-color-neutral5)',
-    neutral10: 'var(--selector-color-neutral10)',
-    neutral20: 'var(--selector-color-neutral20)',
-    neutral30: 'var(--selector-color-neutral30)',
-    neutral40: 'var(--selector-color-neutral40)',
-    neutral50: 'var(--selector-color-neutral50)',
-    neutral60: 'var(--selector-color-neutral60)',
-    neutral70: 'var(--selector-color-neutral70)',
-    neutral80: 'var(--selector-color-neutral80)',
-    neutral90: 'var(--selector-color-neutral90)',
-    primary: 'var(--selector-color-primary)',
-    primary25: 'var(--selector-color-primary25)',
-    primary50: 'var(--selector-color-primary50)',
-    primary75: 'var(--selector-color-primary75)',
-  },
-});
